fix: Resolve 4 pod crash issues — Kafka env vars, data-ingestion profile, confidence column type
- kafka.yaml: Add KAFKA_LOG_DIRS, KAFKA_JMX_PORT, KAFKA_JMX_HOSTNAME to
  fix 'port is deprecated' CrashLoopBackOff on cp-kafka 7.4.0
- data-ingestion.yaml: Switch profile k8s→docker (k8s profile missing),
  add explicit datasource/redis/mqtt env vars for Spring Boot autoconfiguration
- 22-threat-indicators.sql + postgres.yaml ConfigMap: Change confidence
  SMALLINT→INTEGER to match JPA entity (Hibernate validate mode)
- stream-processing will self-resolve once Kafka is healthy

+++ docker/22-threat-indicators.sql
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS threat_indicators (
     indicator_type      VARCHAR(50) DEFAULT 'malicious-activity',
     pattern             TEXT,
     pattern_type        VARCHAR(20) DEFAULT 'stix',
-    confidence          SMALLINT DEFAULT 50 CHECK (confidence BETWEEN 0 AND 100),
+    confidence          INTEGER DEFAULT 50 CHECK (confidence BETWEEN 0 AND 100),
     valid_from          TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     valid_until         TIMESTAMPTZ,
